强制进测试服的开关添加是否是原生设备的判断

+++ assets/Main/Global.js
@@ -51,7 +51,7 @@ window.GlobalCfg = {
   /**
    * 是否强制进测试服
    */
-   is_force_gotoTest: true,
+   is_force_gotoTest: false,
 
   /**
    * 自定义消息分发Id
@@ -1064,8 +1064,12 @@ GlobalCfg.NATIVE_CALL_NAME_OBJ2 = {
 GlobalCfg.IsDownloadPackage2 = 1;
 if (GlobalCfg.IsDownloadPackage2 == 1) {
   GlobalCfg.APP_VERSION = "2.2.2.8";
-  GlobalCfg.APP_INFO_URL = `https://download2.tpgame.in/DownloadPackage2/AppInfo.json`;
-  GlobalCfg.APP_CONFIG_URL = `https://download2.tpgame.in/DownloadPackage2/AppConfig.json`;
+  // GlobalCfg.APP_INFO_URL = `https://download2.tpgame.in/DownloadPackage2/AppInfo.json`;
+  // GlobalCfg.APP_CONFIG_URL = `https://download2.tpgame.in/DownloadPackage2/AppConfig.json`;
+
+
+  GlobalCfg.APP_INFO_URL = `https://download.tkptat.in/production/AppInfo.json`;
+  GlobalCfg.APP_CONFIG_URL = `https://download.tkptat.in/production/AppConfig.json`;
 };
 
 /* 
@@ -1089,7 +1093,7 @@ if (GlobalCfg.isOfflineDeve == 1) {
   * 渠道名：5007
 */
 // GlobalCfg.isOfflineDeve2 = 1;
-if (GlobalCfg.is_force_gotoTest == true) {
+if (GlobalCfg.is_force_gotoTest == true && !cc.sys.isNative) {
   GlobalCfg.IsDownloadPackage2 = 0;
   cc.sys.localStorage.setItem("PackageChannel", "0_7001");
   GlobalCfg.APP_INFO_URL = `https://server.tpmass.com/AppInfo.json?time=${new Date().getTime()}`;
@@ -1388,6 +1392,8 @@ if (packageChannel && packageChannel.indexOf("_") != -1) {
   GlobalCfg.IsDownloadPackage2 = 0;
   let packageChannelArr = packageChannel.split("_"); 
   let server = packageChannelArr[0];
+  console.log("packageChannelArr == " , packageChannelArr)
+  console.log("server == " , server)
   GlobalCfg.server_id = server;
   switch (server) {
     case "0":     // 测试服
@@ -1419,7 +1425,7 @@ if (packageChannel && packageChannel.indexOf("_") != -1) {
       GlobalCfg.APP_CONFIG_URL = `https://download2.tpgame.in/Release4/AppConfig.json`;
       break;
     case "5":     // 5服
-      GlobalCfg.APP_VERSION = "5.0.0.2"; 
+      GlobalCfg.APP_VERSION = "5.0.0.32"; 
       GlobalCfg.APP_INFO_URL = `https://download.tkptat.in/production/AppInfo.json`;
       GlobalCfg.APP_CONFIG_URL = `https://download.tkptat.in/production/AppConfig.json`;
       if (cc.sys.localStorage.getItem("UpdateVersion") == "2.4.13") {//cocos 版本2.4.13
