+++ library/imports/d9/d94231a6-4c4c-43f4-8f88-490f01e94e1a.js
@@ -57,7 +57,7 @@ window.GlobalCfg = (_window$GlobalCfg = {
   /**
    * 是否强制进测试服
    */
-  is_force_gotoTest: true,
+  is_force_gotoTest: false,
   /**
    * 自定义消息分发Id
    */
@@ -990,8 +990,11 @@ GlobalCfg.NATIVE_CALL_NAME_OBJ2 = {
 GlobalCfg.IsDownloadPackage2 = 1;
 if (GlobalCfg.IsDownloadPackage2 == 1) {
   GlobalCfg.APP_VERSION = "2.2.2.8";
-  GlobalCfg.APP_INFO_URL = "https://download2.tpgame.in/DownloadPackage2/AppInfo.json";
-  GlobalCfg.APP_CONFIG_URL = "https://download2.tpgame.in/DownloadPackage2/AppConfig.json";
+  // GlobalCfg.APP_INFO_URL = `https://download2.tpgame.in/DownloadPackage2/AppInfo.json`;
+  // GlobalCfg.APP_CONFIG_URL = `https://download2.tpgame.in/DownloadPackage2/AppConfig.json`;
+
+  GlobalCfg.APP_INFO_URL = "https://download.tkptat.in/production/AppInfo.json";
+  GlobalCfg.APP_CONFIG_URL = "https://download.tkptat.in/production/AppConfig.json";
 }
 ;
 
@@ -1016,7 +1019,7 @@ if (GlobalCfg.isOfflineDeve == 1) {
   * 渠道名：5007
 */
 // GlobalCfg.isOfflineDeve2 = 1;
-if (GlobalCfg.is_force_gotoTest == true) {
+if (GlobalCfg.is_force_gotoTest == true && !cc.sys.isNative) {
   GlobalCfg.IsDownloadPackage2 = 0;
   cc.sys.localStorage.setItem("PackageChannel", "0_7001");
   GlobalCfg.APP_INFO_URL = "https://server.tpmass.com/AppInfo.json?time=" + new Date().getTime();
@@ -1322,6 +1325,8 @@ if (packageChannel && packageChannel.indexOf("_") != -1) {
   GlobalCfg.IsDownloadPackage2 = 0;
   var packageChannelArr = packageChannel.split("_");
   var server = packageChannelArr[0];
+  console.log("packageChannelArr == ", packageChannelArr);
+  console.log("server == ", server);
   GlobalCfg.server_id = server;
   switch (server) {
     case "0":
@@ -1359,7 +1364,7 @@ if (packageChannel && packageChannel.indexOf("_") != -1) {
       break;
     case "5":
       // 5服
-      GlobalCfg.APP_VERSION = "5.0.0.2";
+      GlobalCfg.APP_VERSION = "5.0.0.32";
       GlobalCfg.APP_INFO_URL = "https://download.tkptat.in/production/AppInfo.json";
       GlobalCfg.APP_CONFIG_URL = "https://download.tkptat.in/production/AppConfig.json";
       if (cc.sys.localStorage.getItem("UpdateVersion") == "2.4.13") {
